Add more structure to codebase
