+++ src/handlers/document_save.rs
@@ -1,6 +1,17 @@
 use crate::handlers::RequestHandler;
-use crate::protocol::requests::PolymorphicRequest;
-use crate::shared;
+use crate::protocol::requests::{
+    PolymorphicRequest, Request, TextDocumentSaveParams,
+};
+use crate::shared::create_diagnoistics;
+
+fn parse_save_request(
+    data: String,
+) -> Result<Request<TextDocumentSaveParams>, String> {
+    let request: Request<TextDocumentSaveParams> =
+        Request::from_json(data.as_str())?;
+
+    Ok(request)
+}
 
 #[derive(Default)]
 pub struct DocumentSaveHandler {}
@@ -12,10 +23,10 @@ impl RequestHandler for DocumentSaveHandler {
         prequest: PolymorphicRequest,
         ctx: crate::shared::RequestContext,
     ) -> Result<Option<String>, String> {
-        let request = shared::parse_save_request(prequest.data)?;
+        let request = parse_save_request(prequest.data)?;
         if let Some(params) = request.params {
             let uri = params.text_document.uri;
-            let msg = shared::create_diagnoistics(uri, ctx)?;
+            let msg = create_diagnoistics(uri, ctx)?;
             let json = msg.to_json()?;
 
             return Ok(Some(json));
