+++ src/lib.rs
@@ -2,15 +2,13 @@
 extern crate lazy_static;
 
 pub mod cache;
-pub mod handler;
 pub mod handlers;
 pub mod protocol;
 pub mod shared;
 pub mod stdlib;
-pub mod utils;
 pub mod wasm;
 
 mod visitors;
 
-pub use handler::Handler;
+pub use handlers::Router;
 pub use wasm::Server;
